support flags commands, move to cmd
